requirements.txt
ffmpeg-python==0.2.0
anitopy==2.0.1
fuzzywuzzy[speedup]==0.18.0
toml==0.10.2
PySide6==6.1.1
requests==2.25.1